# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon - OpenAPI 3.0
  description: |-
    Some useful links:
    - [Git](https://github.com/StaZisS/test-task)
  termsOfService: http://swagger.io/terms/
  contact:
    email: inostaziss@gmail.com
  version: 1.0.0
servers:
  - url: https://api.quqee.tech/api/v1 # (если здесь изменить, исправить /ws://localhost:8080 в websocket)
...
paths:
...
  /defects:
    post:
      tags:
        - defect
      summary: Создать заявку с дефектом
      description: Фотограф отправляет снимки с дефектами
      operationId: createDefect()
      security:
        - auth-bearer: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateDefectRequest'


      responses:
        '200':
          description: Информация о заявке.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefectResponseDto'
        '403':
          description: Ошибка при авторизации
    put:
      tags:
        - defect
      summary: Обновить заявку с дефектом
      description: Фотограф обновляет снимки с дефектами полотна
      operationId: updateDefect()
      security:
        - auth-bearer: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateDefectRequest'
      responses:
        '200':
          description: Информация о заявке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefectResponseDto'
        '403':
          description: Ошибка при авторизации
  /defects/{defectId}:
    get:
      tags:
        - defect
      summary: Получить информацию о дефекте
      operationId: getDefect()
      security:
        - auth-bearer: []
      parameters:
        - name: defectId
          in: path
          description: ID заявки, который нужно получить.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Информация о заявке.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalDefectResponseDto'
        '403':
          description: Ошибка при авторизации
...
  /statements:
    post:
      tags:
        - statement
      summary: Создать ведомость с дефектами
      description: Заказчик создает ведомость
      operationId: createStatement()
      security:
        - auth-bearer: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateStatementRequest'
      responses:
        '200':
          description: Созданная ведомость.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatementResponseDto'
        '403':
          description: Ошибка при авторизации
    put:
      tags:
        - statement
      summary: Обновить ведомость с дефектами
      description: Заказчик обновляет ведомость
      operationId: updateStatement()
      security:
        - auth-bearer: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateStatementRequest'
      responses:
        '200':
          description: Обновить ведомость.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatementResponseDto'
        '403':
          description: Ошибка при авторизации
  /statements/{statementId}:
    get:
      tags:
        - statement
      summary: Получить информацию о ведомости
      operationId: getStatement()
      security:
        - auth-bearer: []
      parameters:
        - name: statementId
          in: path
          description: ID отправителя, сообщения которого нужно получить.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Информация о ведомости.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalStatementResponseDto'
        '403':
          description: Ошибка при авторизации
    delete:
      tags:
        - statement
      summary: Удалить ведомость
      operationId: deleteStatement()
      security:
        - auth-bearer: []
      parameters:
        - name: statementId
          in: path
          description: ID ведомости, которую нужно удалить.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '403':
          description: Ошибка при авторизации
...
  /statements/employee:
    post:
      tags:
        - statement
      summary: Назначить на ведомость сотрудника организации
      operationId: addEmployee()
      security:
        - auth-bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentEmployee'
      responses:
        '200':
          description: Информация о ведомости.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatementResponseDto'
        '403':
          description: Ошибка при авторизации
...
components:
  schemas:
...
    CreateDefectRequest:
      type: object
      properties:
        longitude:
          type: number
          format: double
          nullable: false
        latitude:
          type: number
          format: double
          nullable: false
        defect_type:
          type: string
          example: Поперечная трещина
          nullable: false
        defect_distance:
          type: number
          format: double
          nullable: false
        files:
          type: array
          items:
            type: string
            format: binary
    UpdateDefectRequest:
      type: object
      properties:
        defect_id:
          type: string
          format: uuid
          nullable: false
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
        defect_type:
          type: string
          example: Поперечная трещина
        defect_distance:
          type: number
          format: double
        files:
          type: array
          items:
            type: string
            format: binary
    DefectResponseDto:
      type: object
      properties:
        defect_id:
          type: string
          format: uuid
        longitude:
          type: number
          format: double
          nullable: false
        latitude:
          type: number
          format: double
          nullable: false
        defect_type:
          type: string
          example: Поперечная трещина
          nullable: false
        defect_distance:
          type: number
          format: double
          nullable: false
        create_time:
          type: string
          format: date-time
          example: '2025-07-21T17:32:28Z'
          nullable: false
        application_status:
          type: string
          format: string
          example: OPEN
    AdditionalDefectResponseDto:
      type: object
      properties:
        defect_id:
          type: string
          format: uuid
        longitude:
          type: number
          format: double
          nullable: false
        latitude:
          type: number
          format: double
          nullable: false
        defect_type:
          type: string
          example: Поперечная трещина
          nullable: false
        defect_distance:
          type: number
          format: double
          nullable: false
        create_time:
          type: string
          format: date-time
          example: '2025-07-21T17:32:28Z'
          nullable: false
        application_status:
          type: string
          format: string
          example: OPEN
        pictures:
          type: array
          items:
            type: string
            example: https://imgbb.com/picture
    CreateStatementRequest:
      type: object
      properties:
        area_name:
          type: string
          example: Магистраль Москва-Казань
        length:
          type: number
          format: double
        road_type:
          type: string
          example: HIGHWAY
        surface_type:
          type: string
          example: ASPHALT
        direction:
          type: string
          nullable: false
        deadline:
          type: string
          format: date-time
          example: '2025-07-21T17:32:28Z'
          nullable: false
        defects:
          type: array
          items:
            type: string
            format: uuid
        file:
          type: string
          format: binary
    UpdateStatementRequest:
      type: object
      properties:
        area_name:
          type: string
          example: Магистраль Москва-Казань
        length:
          type: number
          format: double
        road_type:
          type: string
          example: HIGHWAY
        surface_type:
          type: string
          example: ASPHALT
        direction:
          type: string
          nullable: false
        deadline:
          type: string
          format: date-time
          example: '2025-07-21T17:32:28Z'
          nullable: false
        statement_status:
          type: string
          example: IN_PROCESS
        organization_performer_id:
          type: string
          format: uuid
        defects:
          type: array
          items:
            type: string
            format: uuid
        file:
          type: string
          format: binary
    AssignmentEmployee:
      type: object
      properties:
        statement_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
    StatementResponseDto:
      type: object
      properties:
        statement_id:
          type: string
          format: uuid
        area_name:
          type: string
          example: Магистраль Москва-Казань
        length:
          type: number
          format: double
        road_type:
          type: string
          example: HIGHWAY
        surface_type:
          type: string
          example: ASPHALT
        direction:
          type: string
          nullable: false
        deadline:
          type: string
          format: date-time
          example: '2025-07-21T17:32:28Z'
          nullable: false
        create_time:
          type: string
          format: date-time
          example: '2024-07-21T17:32:28Z'
          nullable: false
        description:
          type: string
          example: Большие ямы на дорогах
        statement_status:
          type: string
          example: REJECTED
        organization_performer_id:
          type: string
          format: uuid
        organization_creator_id:
          type: string
          format: uuid
          nullable: false
    AdditionalStatementResponseDto:
      type: object
      properties:
        statement_id:
          type: string
          format: uuid
        area_name:
          type: string
          example: Магистраль Москва-Казань
        length:
          type: number
          format: double
        road_type:
          type: string
          example: HIGHWAY
        surface_type:
          type: string
          example: ASPHALT
        direction:
          type: string
          nullable: false
        deadline:
          type: string
          format: date-time
          example: '2025-07-21T17:32:28Z'
          nullable: false
        create_time:
          type: string
          format: date-time
          example: '2024-07-21T17:32:28Z'
          nullable: false
        description:
          type: string
          example: Большие ямы на дорогах
        statement_status:
          type: string
          example: REJECTED
        organization_performer_id:
          type: string
          format: uuid
        organization_creator_id:
          type: string
          format: uuid
          nullable: false
        defects:
          type: array
          items:
            $ref: '#/components/schemas/ShortDefectDto'
...
    ShortDefectDto:
      type: object
      properties:
        defect_id:
          type: string
          format: uuid
        application_status:
          type: string
          example: OPEN
        defect_type:
          type: string
          example: REJECTED
        description:
          type: string
          example: Вожовский переулок
...
  securitySchemes:
    auth-bearer:
      type: http
      scheme: bearer